fix: prevent potential 'nothing to pop' exception on call display state changes
This fix addresses the potential issue of a 'nothing to pop' exception arising when changing the screen call display state,
particularly if this occurs before the completion of the navigation process initiated by a `pushNamed` method call.
The navigator's `pop` method call is now safeguarded with a `canPop` check method call.
Additionally, a new call display state has been implemented to provide listener in the call shell with an extra state change event,
enhancing the execution of `pop` method call.